Added Player class and inherited Tic Tac Toe behaviour
